Add java.util.Date support
Add special support of Date type to manipulate it as internal
representation of epoch in milliseconds. Allow to see the value
without deep reflection (support for jdk17+) and be able
to use it as condition.
Chosen to represent it using epoch to avoid time zone issue:
Date::toString will use the current TimeZone which maybe different
from the user one.