squashing & signing (#8589)
This change introduces TagMap as replacement for HashMap when working with tags.
TagMap has two different implementations...
- one that extends a regular HashMap
- another that uses a different approach that allows for Entry sharing

This change currently uses the HashMap approach by default, but allows for switching to the optimized Map
via a configuration flag.

The optimized TagMap is designed to be good at operations that the tracer performs regularly but HashMap
isn't great at.  Specifically, Map-to-Map copies and storing primitives.

To get the benefit of the optimized TagMap, calling code needs to use TagMap-s for both the source and destination map.
The calling code also needs to make sure to use the bulk operations that which are the most optimized.

To take advantage of TagMap in span creation, a mechanism was introduced that allows for bypassing TagInterceptors.
Bypassing TagInterceptors is done by analyzing a TagMap that is going to be reused ahead-of-time to determine if
interception is needed.  If interception isn't needed, then SpanBuilder can use a bulk operation to update the map.

To maintain the insertion order semantics of SpanBuilder, TagMap also includes a Ledger.
A Ledger is concatenative ledger of entry modifications to a map
A Ledger is now used in place of a LinkedHashMap in SpanBuilder to provide the insertion order semantics.

Since these changes primarily serve to reduce allocation, they should the biggest gain in memory constrained environments.
In memory constrained environments, these changes yield a 10% increase in sustainable throughput with spring-petclinic.