Prevent crash in SQL Server JDBC when tracing execute methods with generated keys
When tracing the JDBC statement execute methods, we prepend a comment to
the SQL query used for DBM trace propagation.
However, there is a bug in the SQL Server JDBC driver that prevents the
generated keys from being returned when the SQL comment is prepended to
the SQL query.
I decided to only append the comment in this specific case to avoid the
comment from being truncated.

@see https://github.com/microsoft/mssql-jdbc/issues/2729