Add peer tags, span kind and trace root flag to MetricKey bucket (#9178)
* chore(css): Add peer tags, span kind and trace root flag to MetricKey bucket

* chore(css): Use a map to create the metric key

Also delays the creation of the Utf8ByteString at serialization time.
Note that `writer.writeUTF8` emits a header corresponding to the length
of the value being written, calling this method again will count as an
entry in the array.
A possible idea was to use a cache to store the computation of the
peerTags, however `Utf8ByteString` is not concatenable / appendable,
which is necessary to have the proper encoding. Creating a
new "temporary" `Utf8ByteString`, was replaced by a direct call to
`String: :getBytes`.

* Improves CSS peer tag aggregation (#9336)

* Add jmh for metrics aggregation

* Cache peer tags to avoid too many strings/utf8 conversions

* Use span kind cache

* Fix tests

* Hardcode eligible span kind since  agent backpropagated are deprecated

* revisit peer tags aggregation rules according to the rfc

* IsTraceRoot is a tristate

* Don't confuse trace root with top levels

* Fix build after rebase

---------

Co-authored-by: Andrea Marziali <andrea.marziali@datadoghq.com>