Send only snapshots to Debugger Track
We introduce a separate BatchUploader instance to send
capureSnapshot=true snapshots (including variables) to dedicated
datadog endpoint for DEBUGGER track.
SnapshotSink now expect 2 BatchUploaders (lowRate for snapshots and
highRate for templated logs only)
BatchUploader now take a name to identify in debug logs which instance
is used.
Smoke tests are updated to use processRequest everywhere and get rid of
retrieveSnapshots that is not correct.