Add Capture Expressions
Previously watches, capture expressions allow to describe what exactly
 we want to capture and put into a snapshot for a log probe.
When one or more capture expressions is defined, only the capture
expressions are added to the Captures part of the snapshot.
Expressions can be complex using the expression language and must
return a value from which we start to capture the object graph
following the limits that can be defined per capture expressions.
A new attribute captureExpressions is added into the snapshot.