Simplify type outlines to not implement type.getDeclaringType() and type.getEnclosingType()
These methods are not used for matching purposes, so don't need to be maintained in outlines.
There is one use of type.isAnonymousType() in IAST, but this use is under a feature flag that
is off by default. To satisfy that use we provide a simple isAnonymousType() implementation
which is based on the '$number' convention for anonymous class names.

Note: if anything does touch type.getDeclaringType() / type.getEnclosingType() later on when
doing the final transformation then the outline type is automatically inflated to a full type.
(This is already the case for other methods that aren't available in type outlines.)

We also drop type.getClassFileVersion() from outlines for the same reason.