chore: Adds a test for JDK-827271
After JDK 18/19 javac compiles differently `Object` methods, while there is a ignore mechanism on java.* packages
after the key change is described in https://bugs.openjdk.org/browse/JDK-827271, `Object`'s methods
are applied to interfaces which may belong to another package.

Key change of JDK-827271
"invocations of `java.lang.Object` methods on interfaces in the classfile will use `invokeinterface` referring
to the interface, which is consistent with JLS 9.2. This will be done regardless of whether the interface
declares the method explicitly or not."