Adds a test for JDK-827271 (#9650)
* Ignore java.lang.Object methods (identified by method name plus method descriptor) when creating muzzle references.

This helps when compiling with Java 18+ because javac will now use INVOKEINTERFACE with the interface name to call
java.lang.Object methods (so the 'owner' is now seen as the interface) whereas before it was using INVOKEVIRTUAL
with java.lang.Object as the method 'owner'.

the key change is described in https://bugs.openjdk.org/browse/JDK-8272715:
  "invocations of java.lang.Object methods on interfaces in the classfile will use invokeinterface referring
   to the interface, which is consistent with JLS 9.2. This will be done regardless of whether the interface
   declares the method explicitly or not."

* chore: Adds a test for JDK-827271

After JDK 18/19 javac compiles differently `Object` methods, while there is a ignore mechanism on java.* packages
after the key change is described in https://bugs.openjdk.org/browse/JDK-827271, `Object`'s methods
are applied to interfaces which may belong to another package.

Key change of JDK-827271
"invocations of `java.lang.Object` methods on interfaces in the classfile will use `invokeinterface` referring
to the interface, which is consistent with JLS 9.2. This will be done regardless of whether the interface
declares the method explicitly or not."

---------

Co-authored-by: Stuart McCulloch <stuart.mcculloch@datadoghq.com>