Simplify type outlines to not implement type.getDeclaringType() and type.getEnclosingType() (#9644)
These methods are not used for matching purposes, so don't need to be maintained in outlines.

There is one use of type.isAnonymousType() in IAST, but this is under a support flag - I could not find any active use of that support flag. To satisfy the original use-case (involving MyBatis) we add a simple anonymous type heuristic to that particular instrumentation, based on the '$number' Java language convention for anonymous class names.

( If necessary we can further enhance that specialized matcher without incurring overhead for all outline parsing )

Note: if anything does touch type.getDeclaringType() / type.getEnclosingType() later on when doing the final transformation then the outline type is automatically inflated to a full type. (This is already the case for other methods that aren't available in type outlines.)

We also drop type.getClassFileVersion() from outlines for the same reason.