Opt-out API Security endpoint discovery (#9623)
We were keeping this feature disabled for now because the backend does not support it yet. However, we’re actively working on adding support, and in the meantime, we need to measure the volume of data we’d need to handle once the feature is enabled.

As a result, this PR sets the default config flag to true:
ConfigDefaults.java#L120

Enabling the feature by default surfaced an issue during test execution:
an exception was thrown due to multiple beans of type RequestMappingHandlerMapping being present in the Spring context, which caused a NoUniqueBeanDefinitionException.

Error:

No qualifying bean of type 'RequestMappingHandlerMapping' available: 
expected single matching bean but found 2: requestMappingHandlerMapping,controllerEndpointHandlerMapping
To resolve this, the instrumentations were updated to:

Use getBeansOfType(...) instead of getBean(...)
Iterate over all available RequestMappingHandlerMapping beans
Collect and merge their handler methods safely.
This avoids the error and ensures compatibility with Spring setups that register multiple handler mappings.