Remove the warning on kotlin plugin warning (#9711)
* fix: Kotlin plugin warning

Build scripts use the `buildscript` way to configure the project's plugins, which is not recommended anymore since quite a few versions in gradle.

```
The Kotlin Gradle plugin was loaded multiple times in different subprojects, which is not supported and may break the build.
This might happen in subprojects that apply the Kotlin plugins with the Gradle 'plugins { ... }' DSL if they specify explicit versions, even if the versions are equal.
Please add the Kotlin plugin to the common parent project or the root project, then remove the versions in the subprojects.
If the parent project does not need the plugin, add 'apply false' to the plugin line.
See: https://docs.gradle.org/current/userguide/plugins.html#sec:subprojects_plugins_dsl
The Kotlin plugin was loaded in the following projects: ':dd-java-agent:agent-ci-visibility', ':dd-java-agent:agent-llmobs'
```

* style: make scalafmt happy

* style: make google-java-format happy

* fix: Exclude generated files and more generally anything in the build directory

* fix: Single kotlin plugin version but configure each projects