Fix GraalVM smoke test on Java 21
The root issue was that the reference to MemFDUnixWriter in the TracerInstaller
bytecode was being picked up by HelperScanner, which is a utility that we use to
avoid having to maintain an exact list of classes to inject into GraalVM in order
to activate the tracer in the final native binary via VMRuntimeInstrumentation
and TracerActivation.

The simplest solution was to use reflection to load MemFDUnixWriter.

The alternative would have been to enhance HelperScanner to be able to exclude
types from the list of classes it finds. However this would be more complicated.