optimize line probe without condition
In case of a line probe without condition we can move the sampling is
currently done in LogProbe::evaluate to LogProbe::isReadToCapture.
This way if the sample fails we return false on isReadyTocapture
and no CapturedContext is created and nothing captured.
If we sample the execution the CapturedContext will be created
we capture the context and seiralize the result, but we are sure that
we are not doing for nothing.