Add Probe Metadata (#9723)
Introducing ProbeMetadata class to help get probe from instrumented
 code. Instead of getting the probe instance from a map based on
probeId we are directly getting it from an array that is filled at
instrumentation time. This way it's cheaper and abstracted the way to
get probe definition to evaluate conditions for example.
In instrumented code, instead of referring the probe by the probeId
hard-coded we are using index in probe metadata array.
ProbeMetadata class encapsulate an AtomicReferenceArray to make sure
every cell as a volatile-style access and provide linear way to remove
or add into this array.
For now linear scanning is fine as we don't expect a lot of probes in
a same location. Could improve on that using bitmaps later if require.
Instrumentation is changed to push int arrays instead of String arrays
and can be optimize later to accept static arguments instead of
varargs/array instead.