Split CapturedContext instrumenters (#9778)
make it a specialized version for single probe and keep the previous
one as the fallback when multiple probes need to be instrumented for
the same location.
Introduced SingleCapturedContextIntrumenter and
MultipleCapturedContextInstrumenter based on current
CapturedContextInstrumenter. Introduced specialized version with only
one probe index of methods in DebuggerContext called from the
instrumetnation. Introduced CapturedContextProbe interface and move
some specific methods for LogProbe from ProbeImplementation.
This allow to implement later some optimized version of the
instrumentation which is not the case right now. This for now just a
refactoring for preparing the next ones