support service discovery with JNA (#9705)
* try to plug memfd to core-tracer

* prepare encoding

* add proper encoding

* suggestions

* add groovy test

* spotless

* fix graalvm runs

* fix test

* try something for graalvm

* Fix GraalVM smoke test on Java 21

The root issue was that the reference to MemFDUnixWriter in the TracerInstaller
bytecode was being picked up by HelperScanner, which is a utility that we use to
avoid having to maintain an exact list of classes to inject into GraalVM in order
to activate the tracer in the final native binary via VMRuntimeInstrumentation
and TracerActivation.

The simplest solution was to use reflection to load MemFDUnixWriter.

The alternative would have been to enhance HelperScanner to be able to exclude
types from the list of classes it finds. However this would be more complicated.

* Skip reflection call when launching native image

* Suppress warning about use of Class.forName because it's intentional

* Lazily load and execute service discovery

* hide subprocess span due to jna command execution

* update after review

* match json spec for python decoder in system-tests

* update after review

* add config to disable

* Rearrange config

* update after review

* update after review

* Apply suggestions from code review

Co-authored-by: Stuart McCulloch <stuart.mcculloch@datadoghq.com>

* feat: Refactor factory and introduce NOOP

* feat: Removing noop to avoid always schedule

* match agent prefix matcher

---------

Co-authored-by: Andrea Marziali <andrea.marziali@datadoghq.com>
Co-authored-by: Stuart McCulloch <stuart.mcculloch@datadoghq.com>
Co-authored-by: Bruce Bujon <bruce.bujon@datadoghq.com>