optimize method probe without condition (#9834)
In case of a method probe without condition we can move the sampling
is currently done in LogProbe::evaluate to LogProbe::isReadToCapture.
This way if the sample fails we return false on isReadyTocapture and
no CapturedContext is created and nothing captured. If we sample the
execution the CapturedContext will be created we capture the context
and serialize the result, but we are sure that we are not doing for
nothing.
We introduce also a differentiation for single probe/multiple probe
in calling evaluate that is propagated to probe implementation. it
helps to differentiate the behavior based on the instrumentation