Track long running traces when agent does not support long running feature
This allows dumping long running traces when not connected to a
Datadog Agent using the new JMX flare feature. A warning message
will be logged in this case to indicate that long running traces
will not be sent upstream but are available in a flare.

Previously the long running traces buffer would always be empty,
even though the feature was enabled with
dd.trace.experimental.long-running.enabled=true. This led to a
good amount of confusion when I was initially developing a feature
to dump long running traces without a local Datadog Agent running.