Fix logging when uninstrumenting cannot happen (#9851)
if we have a thread in a loop that cannot exit the instrumented method
for a while, instrumentation cannot be done. But probe is removed
from metadata and cannot be resolved. We were logging this status
which is in fact not helpful and generate noise.
If probe cannot be resolved we just ignore the instrumentation