Add long-running traces metric for drops due to sampling priority
This likely isn't an important metric to track, but I noticed these
long-running traces were the only ones not reflected in existing
metrics when they are dropped from the tracker, so I thought it might
be good to add this metric for completeness.

This change introduces a new metric, "long-running.dropped_sampling",
to count traces that are dropped when negativeOrNullPriority(trace)
is true.

There is an existing metric, "long-running.dropped", for long-running
traces that are dropped on input to the tracker when there are no
slots free. That metric name was kept as-is to not disturb any
existing consumers downstream. If that is not a concern, it might
be good to rename the existing metric to clarify that it captures
traces dropped on input.