fix JDBC's SQLCommenter not taking into account semicolons
After adding the new always append comment feature in #9798, a customer
found a bug that was present in our SQL Commenter since a long time.
Basically, JDBC always split queries to send to the DB on the semicolon.
If a comment is after a semicolon, JDBC will still split the queries,
and return multiple ResultSet. This provoked exceptions like this one:
```
org.postgresql.util.PSQLException: Multiple ResultSets were returned by
the query
```

SDBM-2100