Add instrumentation error details on test failures
This puts the details front and center in the test summary and is
much easier than hunting back through log messages.

Note: In InstrumentationErrors, errors are only recorded when
enableRecordingAndReset has been called. This is only done from the
InstrumentationSpecification test specification, so we don't need
to worry about the ArrayList growing without bounds in production.
I considered adding a limit to the ArrayList, but opted not to for
simplicity. I also wanted to avoid silently discarding some errors.