Add App Logs Collection feature (#10156)
Add application logs collection based on log framework instrumentation
that hooks when log event are generated to be written. Log events are
sent to Logs backend through Datadog agent's Event Platform (Evp)
proxy. We are reusing a previous infrastructure used for uploading
logs for CI Visibility (see #7082).
We are adding support for Logback.
This feature is opt-in through DD_APP_LOGS_COLLECTION_ENABLED config.
Logs are sent as structured log layout and allow to add any additional
attributes. For now we are adding directly trace correlation ids
(traceId and spanId) and give the logs injection feature for free and
without any setup on the log framework side.
Smoke tests are based on Logs Injection tests.